Improve RSS feed: Switch to startDateIso8601 and include relevant information
Removes 'pubDate' tag, as this is commonly used for specifying the date
when the feed entry got updated. I could not find a jekyll-way of
publishing that information.

Switched from using 'datestring' to 'startDateIso8601' as Jekyll could
e.g. not parse datestring of TP77. Furthermore, included location in the
title of the feed.

Changed the description from (non-existing) 'content' to list of
speakers + talk title.